lab 3 cu clasele din domain + repo

+++ lab3/src/main/java/tudor/domain/Client.java
@@ -1,6 +1,6 @@
 package tudor.domain;
 
-public class Client extends Entity<Long>{
+public class Client extends Entity<Integer>{
     String name;
     String addres;
 }
