+++ lab3/src/main/java/tudor/domain/Match.java
@@ -1,6 +1,6 @@
 package tudor.domain;
 
-public class Match extends Entity<Long>{
+public class Match extends Entity<Integer>{
     String team1;
     String team2;
     String matchType;
