+++ lab3/src/main/java/tudor/domain/Ticket.java
@@ -1,6 +1,6 @@
 package tudor.domain;
 
-public class Ticket extends Entity<Long> {
+public class Ticket extends Entity<Integer> {
     Client client;
     Match match;
     int numberOfSeats;
